BIP1: clarify early stages of BIP championing
The previous BIP1 text was ambiguous regarding early steps for taking an
idea from concept and eventually into a BIP. The new text is intended to
make it more clear that the initial email to the bitcoin-dev mailing
list should not be a fully-formed BIP. There have been exceptions to
this in the past for ideas already widely known and implementation in
progress, but "you know it when you see it". Hopefully this will add
clarity to the BIP authoring process and work flow for new authors.